pyproject.toml
[tool.poetry.scripts]
cli = "cli"
web = "diet.main:main"

[tool.poetry]
name = "diet"
version = "1.0.0"
description = "Let's diet!"
authors = ["23prime <23.prime.37@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^0.25.3"
matplotlib = "^3.1.2"
flask = "^1.1.1"
SQLAlchemy = "^1.3.12"
psycopg2 = "^2.8.4"
pyopenssl = "^19.1.0"
python-twitter = "^3.5"
flask-sqlalchemy = "^2.4.1"
flask-wtf = "^0.14.2"
flask-bootstrap = "^3.3.7"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5"
hacking = "^2.0.0"
isort = "^4.3.21"
yapf = "^0.29.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
